+++ src/Index.ts
@@ -1,7 +1,9 @@
-import { initGithub } from './GitHub';
-import { initGitLab } from './GitLab';
-import { initBitBucket } from './BitBucket';
+import { initGithub } from './pages/GitHub';
+import { initGitLab } from './pages/GitLab';
+import { initBitBucket } from './pages/BitBucket';
+import { initGistGithub } from './pages/GistGitHub';
 
 initGithub();
 initGitLab();
 initBitBucket();
+initGistGithub();